+++ History.md
@@ -3,6 +3,7 @@
 1.1.2
 =====
 * Make sure anyhow (logger) is set up on init.
+* Calling reset() will deep clean instead of create new settings object.
 
 1.1.1
 =====